refactor(spec): extract builder scale sugar and geom options
Split the fluent builder into core, scale mixin, and geom option types so
thin scale wrappers and geom option surfaces evolve independently. Co-locate
position-scale schema vs helper/builder tests. Public gg()/GGBuilder API is
unchanged.
