+++ .changeset/spec-builder-scales-split.md
@@ -1,0 +1,7 @@
+---
+"@ggsvelte/spec": patch
+---
+
+<!-- markdownlint-disable MD041 -->
+
+Extract fluent builder geom option types and scale sugar into focused modules, and split position-scale schema vs helper tests. Public gg()/GGBuilder API is unchanged.
